Russian translation added.
Hello!
I am the one who have made a complete overhaul of official Russian TTT translation! And my friend (dok441) said that there is a TTT2 that haven't been translated yet. So, here is my translation of TTT2. (many strings are the same as in official Russian TTT translation and it's obvious why)